Modified Docker compose, now the following services are runnable in docker: Salon Review and User
They require a little bit of configuration: A sql initalization script has been added to initialize the databases we need for the microservices using mysql.

In the docker compose there environement variables, so you need to create a .env file at the root repo and assign a value to the variables. For the mongodb connection uri and neo4j, I will send them in the discord group. For the rest, you need to choose a username and password for the mysql db connection, the same that you will use in application.proper
of every the microservices using mysql.